+++ proto/CHANGELOG.md
@@ -23,6 +23,8 @@ To release a new version (e.g. from `1.0.0` -> `2.0.0`):
 
 ## [Unreleased]
 
+## [1.3.1] - 2026-09-01
+
 *   Update EDA outcome severity documentation in README_internal.md.
 *   Add `knots_spec` oneof (`n_knots` and `knot_locations`) to `Hyperparameters`
     proto to disambiguate single-element knots serialization and
@@ -77,4 +79,5 @@ To release a new version (e.g. from `1.0.0` -> `2.0.0`):
 [1.2.1]: https://github.com/google/meridian/releases/tag/proto-v1.2.1
 [1.2.2]: https://github.com/google/meridian/releases/tag/proto-v1.2.2
 [1.3.0]: https://github.com/google/meridian/releases/tag/proto-v1.3.0
-[Unreleased]: https://github.com/google/meridian/compare/proto-v1.3.0...HEAD
+[1.3.1]: https://github.com/google/meridian/releases/tag/proto-v1.3.1
+[Unreleased]: https://github.com/google/meridian/compare/proto-v1.3.1...HEAD